reports/paloalto/requirements.txt
pytest==7.4.0
pytz
xmltodict
pan-os-python>=1.11.0
azure-identity
azure-keyvault-keys
azure-keyvault-secrets
azure-mgmt-compute>=17.0.0
azure-mgmt-keyvault>=7.0.0
azure-mgmt-network>=16.0.0
azure-mgmt-resource>=15.0.0
azure-mgmt-core>=1.4.0
azure-cosmos>=4.4.0
